+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.ui.order.OrderDetailActivity"
+            android:name=".presentation.ui.order.detail.OrderDetailActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.ui.myPage.MyPageActivity"
