+++ app/src/main/java/woowacourse/shopping/presentation/ui/order/detail/OrderDetailPresenter.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.ui.order
+package woowacourse.shopping.presentation.ui.order.detail
 
 import woowacourse.shopping.domain.repository.OrderRepository
 import woowacourse.shopping.domain.util.WoowaResult
